Completed MFG Eff Si Calculations
ALL DONE! Potentially minor tweaks if errors are found, but a mfg eff for Si has been calculated, it outputs a csv, and there is a pretty graph at the end to compare the rise of the diamond saw to the efficiency of silicon usage. There is a slight step change. The results have been copied into the silicon material baseline file.