Updated "reduce capacity"
reduce capacity is an input in the calculate mass flow function, which reduces capacity installed by the bifaciality factor. This is now set to "False" by default, i.e. it will only install by nameplate capacity, and bifacial irradiance is bonus. This will not affect the energy generation calculation, which includes bifaciality (in irradiance).