change mfging scrap HQ energy to additional energy
To avoid double counting energy demands in the aggregateEnergy function, the LQ energy needs to be removed from this line. This implies that the HQ is "additional energy required" not the total energy of getting something to HQ. This can be updated in future with larger changes to the maths.